+++ src/shared/ui/button/login-button.stories.ts
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from "@storybook/react";
-import { Button } from "./button";
+import { Button } from "./login-button";
 
 const meta: Meta<typeof Button> = {
   title: "Shared/UI/Button",
